refactor: Payment totalCost 타입 변경

+++ src/main/java/com/idukbaduk/itseats/payment/dto/PaymentInfoRequest.java
@@ -15,7 +15,7 @@ public class PaymentInfoRequest {
 
     private Long orderId;
     private List<Long> coupons;
-    private int totalCost;
+    private Long totalCost;
     private String paymentMethod;
     private String storeRequest;
     private String riderRequest;
